Add Feedback Popup plugin and update version
A feedback popup plugin was added to enhance the user interaction with the site. The plugin id was then added in the list of plugins for the NextJsSupabaseLite model. The version of the codebase was updated to reflect these changes in 'package-lock.json'. Instructions for using the CLI were also updated in 'README.md' for greater clarity. Also, a '.npmrc' file was added to specify the registry URL for the npm package '@makerkit'.