Update blogPath and plugins for remix-supabase
Adjusted the blogPath to `app/content/posts` from an empty string to resolve the error of not being able to correctly route to the blog content for the remix-supabase kit.

Also added AiChatBot, FeedbackPopup, and CookieBanner plugins into the `plugins` array for the remix-supabase kit to enhance its features.

Additionally, bumped up the version of the Makerkit CLI to 1.2.6.